+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- New resize feature to `QasUploader` (auto resize).
+- `QasUploader` new props: (acceptResizeTypes, picaResizeOptions, sizeLimit, useResize).
+
+### Changed
+- Updated `vue` and `vue-template-compiler` to `v2.6.14` for fixing storybook error.
+
 ## 2.9.6 - 2021-09-22
 
 ### Added